Updated How to choose unique EPICS from GAIA Data
Data from GAIA contained multiple rows of data for the same EPIC. Originally,
we chose the row with the most amount of entries. Althought this may actually
correspond to the row we want, this is not justifiable. Now, the program
chooses the row of data which has the minimum GAIA absolute magnitude, i.e.
the brightest (and hopefully most accurate) data point.