Updated SOM Work
The SOM bins are now normalised between 0 and 1 so that the min occurs at (0, 0)
and the max value is 1.0. Some statistics are still needed to be generated on
these binned values. Future work will involve printing the kohonen layer.

Noise and OTHPER have been down-sampled to roughly 100 per som so that they
do not dominate.

 All classes are now plotted.